remove old code + small fixes
