+++ app/layout.tsx
@@ -1,25 +1,25 @@
-import Link from 'next/link'
-import React from 'react'
+import Link from "next/link";
+import React from "react";
 
 export default function RootLayout({
   children,
+}: {
+  children: React.ReactNode;
 }) {
   return (
     <html lang="en">
-      <body style={{
-        margin: '3rem',
-      }}>        
+      <body
+        style={{
+          margin: "3rem",
+        }}
+      >
         <header>
-          <Link href="/">
-            Home
-          </Link>
-          {' | '}
-          <Link href="/posts">
-            Posts
-          </Link>
+          <Link href="/">Home</Link>
+          {" | "}
+          <Link href="/posts">Posts</Link>
         </header>
         <main>{children}</main>
       </body>
     </html>
-  )
+  );
 }
