+++ app/posts/page.tsx
@@ -1,11 +1,8 @@
-import PostList from "./client-page";
+import PostList from "./post-list";
 import { client } from "../../tina/__generated__/client";
 
-
-
 export default async function Page() {
-
   const pages = await client.queries.postConnection();
 
-  return <PostList {...pages} />
+  return <PostList {...pages} />;
 }
